+++ app/contacts/do-contacts-recDuplicate.php
@@ -3,8 +3,8 @@
 $ValidationMsgA = array();
 
 // Validate Contact ID
-if ( \xan\isEmpty( $doParam[ 'IDContacts' ] ) ) {
-    $ValidationMsgA[] = "Contact ID is Blank";
+if ( \xan\isEmpty( $doParam[ $mmContactsT->NameTableParam ] ) ) {
+    $ValidationMsgA[] = $mmContactsT->NameSingular . ' ID is Blank';
 }
 
 // Validate Response
@@ -19,11 +19,11 @@ $UUIDNew = "";
 
 // Duplicate Record
 $recs = new \xan\recs( $mmContactsT );
-$recs->recordDuplicate( $doParam[ 'IDContacts' ] );
+$recs->recordDuplicate( $doParam[ $mmContactsT->NameTableParam ] );
 
 // Error Check
 if ( $recs->errorB || $recs->rowCount < 1 ) {
-	$ValidationMsgA[] = 'Contact Duplicate Error' . $recs->messageExtra . '; ' . $recs->messageSQL;
+	$ValidationMsgA[] = $mmContactsT->NameSingular . ' Duplicate Error' . $recs->messageExtra . '; ' . $recs->messageSQL;
 } else {
 	// Recs Loop
 	$recs->rowIndex = -1;
@@ -45,11 +45,11 @@ if ( !empty( $ValidationMsgA ) ) {
 
 // Duplicate Related ContactComms
 $recs = new \xan\recs( $mmContactsCommsT );
-$recs->recordDuplicateRelated( $mmContactsT->NameTableKey, $doParam[ 'IDContacts' ], $UUIDNew );
+$recs->recordDuplicateRelated( $mmContactsT->NameTableKey, $doParam[ $mmContactsT->NameTableParam ], $UUIDNew );
 
 // Error Check
 if ( $recs->errorB ) {
-	$ValidationMsgA[] = 'Comms Duplicate Error' . $recs->messageExtra . '; ' . $recs->messageSQL;
+	$ValidationMsgA[] = $mmContactsCommsT->NameSingular . ' Duplicate Error' . $recs->messageExtra . '; ' . $recs->messageSQL;
 }
 
 // Validate Response